Join tests files in just one
